refactor: 公開された submit 宣言の読み出しを 1 か所にする
jscpd がプレビュー側と参加側で同じ 58 トークンを見つけた。同じ文書
（config/public.submit[cid]）を、書かれる前と書かれた後から読んでいる
二重の綴りだったので submitSpec.ts に出した。
